+++ src/main/java/dev/fulmineo/guild/data/QuestHelper.java
@@ -145,14 +145,18 @@ public class QuestHelper {
 	}
 
 	public static NbtCompound writeProfessionData(NbtCompound nbt, GuildServerPlayerEntity guildPlayer, QuestProfession profession) {
-		List<QuestLevel> levels = DataManager.levels.get(profession.levelsPool);
+		List<QuestLevel> levels = ServerDataManager.levels.get(profession.levelsPool);
 		int exp = guildPlayer.getProfessionExp(profession.name);
 		int level = QuestHelper.getCurrentLevel(levels, exp);
 		nbt.putInt("Level", level);
-		QuestLevel currentLevel = levels.get(level);
-		QuestLevel nextLevel = levels.get(level+1);
-		nbt.putInt("LevelPerc", (int)(((float)(exp - currentLevel.exp) / (float)(nextLevel.exp - currentLevel.exp)) * 100));
-		nbt.putBoolean("LevelMax", level == levels.size() - 1);
+		boolean levelMax = level == levels.size() - 1;
+		if (levelMax) {
+			nbt.putInt("LevelPerc", 100);
+		} else {
+			QuestLevel currentLevel = levels.get(level);
+			QuestLevel nextLevel = levels.get(level+1);
+			nbt.putInt("LevelPerc", (int)(((float)(exp - currentLevel.exp) / (float)(nextLevel.exp - currentLevel.exp)) * 100));
+		}
 		return nbt;
 	}
 }