+++ src/main/java/dev/fulmineo/guild/data/QuestProfessionRequirement.java
@@ -1,0 +1,6 @@
+package dev.fulmineo.guild.data;
+
+public class QuestProfessionRequirement {
+	public String profession;
+	public Range level;
+}