+++ src/main/java/dev/fulmineo/guild/data/Range.java
@@ -6,20 +6,15 @@ public class Range {
 	public Integer min;
 	public Integer max;
 
-	private void init() {
-		if (this.min == null) this.min = 1;
-		if (this.max == null) this.max = 64;
-	}
-
 	public boolean contains(int value) {
-		this.init();
-		return value >= this.min && value <= this.max;
+		return (this.min == null || value >= this.min) && (this.max == null || value <= this.max);
 	}
 
 	public int getQuantityInRange() {
-		this.init();
-		int diff = max - min;
-		int val = (diff > 0 ? (new Random()).nextInt(diff + 1) : 0) + min;
+		int qmin = this.min == null ? 1 : this.min;
+		int qmax = this.max == null ? 64 : this.max;
+		int diff = qmax - qmin;
+		int val = (diff > 0 ? (new Random()).nextInt(diff + 1) : 0) + qmin;
 		return val;
 	}
 }